Add 'spin-serde' crate
This collects some serde tools into a separate crate, avoiding some
inter-crate dependencies.
