+++ crates/app/Cargo.toml
@@ -7,9 +7,9 @@ edition = { workspace = true }
 [dependencies]
 anyhow = "1.0"
 async-trait = "0.1"
-base64 = "0.21.3"
 ouroboros = "0.18.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 spin-core = { path = "../core" }
+spin-serde = { path = "../serde" }
 thiserror = "1.0"
