+++ crates/serde/src/lib.rs
@@ -1,0 +1,15 @@
+//! Serde serialization helpers
+
+#![deny(missing_docs)]
+
+pub mod base64;
+pub mod id;
+mod version;
+
+pub use version::{FixedStringVersion, FixedVersion};
+
+/// A "kebab-case" identifier.
+pub type KebabId = id::Id<'-'>;
+
+/// A "snake_case" identifier.
+pub type SnakeId = id::Id<'_'>;
